fix: clean up
* rename (integration) tests around (Mongo)Schedule
* add test for MongoSchedule.connect and disconnect
*do not log before user has had a chance to listen to emitted log events
